main: awb: Provide an allocation
Provide an allocation of the data structure, and assign it
to the pointer which is passed to ia_aiq_awb_run().

This doesn't work - and awb_result still gets set to NULL.

This could likely be ia_aiq_awb_run() reporting an error, as it's
a bit odd that the function would modify the storage pointer, or in
fact take a pointer to a pointer at all.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>